complete auth rework
