+++ src/store/auth.js
@@ -1,5 +1,4 @@
 import { createSlice } from '@reduxjs/toolkit';
-import getCookieValue from '../util/getCookieValue';
 
 const initialAuthState = { token: '' };
 
@@ -7,45 +6,8 @@ const authSlice = createSlice({
   name: 'auth',
   initialState: initialAuthState,
   reducers: {
-    checkTokenInCookie(state) {
-      let tokenCookieValue = getCookieValue('bearerToken');
-
-      if (!tokenCookieValue) {
-        authSlice.caseReducers.getBearerToken(state);
-      } else {
-        state.token = tokenCookieValue;
-      }
-    },
-
-    async getBearerToken(state) {
-      const clientId = process.env.REACT_APP_CLIENT_ID;
-      const key = process.env.REACT_APP_KEY;
-
-      const response = await fetch('https://api.petfinder.com/v2/oauth2/token', {
-        body: `grant_type=client_credentials&client_id=${clientId}&client_secret=${key}`,
-        headers: {
-          'Content-Type': 'application/x-www-form-urlencoded',
-        },
-        method: 'POST',
-      });
-
-      if (!response.ok) {
-        throw Error('Something went wrong with API authentication!');
-      }
-
-      const data = await response.json();
-      const setTokenAction = { type: 'setToken', token: data.access_token, time: data.expires_in };
-
-      authSlice.caseReducers.setBearerToken(state, setTokenAction);
-    },
-
     setBearerToken(state, action) {
-      const dateObj = new Date();
-      const dateTime = dateObj.getTime();
-      const expiry = new Date(dateTime + action.time * 1000);
-
-      document.cookie = `bearerToken=${action.token};expires=${expiry}};path=/`;
-      state.token = action.token;
+      state.token = action.payload.token;
     },
   },
 });
